hci_transport_hX_posix: extract set baudrate into btstack_uart_posix
