+++ platform/posix/hci_transport_h4_posix.c
@@ -55,6 +55,7 @@
 #include "btstack_debug.h"
 #include "hci.h"
 #include "hci_transport.h"
+#include "btstack_uart_posix.h"
 
 #ifdef HAVE_EHCILL
 #error "HCI Transport H4 POSIX does not support eHCILL yet. Please remove HAVE_EHCILL from your btstack-config.h"
@@ -106,63 +107,9 @@ static uint8_t hci_packet_with_pre_buffer[HCI_INCOMING_PRE_BUFFER_SIZE + 1 + HCI
 static uint8_t * hci_packet = &hci_packet_with_pre_buffer[HCI_INCOMING_PRE_BUFFER_SIZE];
 
 static int    h4_set_baudrate(uint32_t baudrate){
-
     log_info("h4_set_baudrate %u", baudrate);
-
-    struct termios toptions;
     int fd = btstack_run_loop_get_data_source_fd(hci_transport_h4->ds);
-
-    if (tcgetattr(fd, &toptions) < 0) {
-        perror("init_serialport: Couldn't get term attributes");
-        return -1;
-    }
-    
-    speed_t brate = baudrate; // let you override switch below if needed
-    switch(baudrate) {
-        case 57600:  brate=B57600;  break;
-        case 115200: brate=B115200; break;
-#ifdef B230400
-        case 230400: brate=B230400; break;
-#endif
-#ifdef B460800
-        case 460800: brate=B460800; break;
-#endif
-#ifdef B921600
-        case 921600: brate=B921600; break;
-#endif
-
-// Hacks to switch to 2/3 mbps on FTDI FT232 chipsets
-// requires special config in Info.plist or Registry
-        case 2000000: 
-#if defined(HAVE_POSIX_B300_MAPPED_TO_2000000)
-            log_info("hci_transport_posix: using B300 for 2 mbps");
-            brate=B300; 
-#elif defined(HAVE_POSIX_B1200_MAPPED_TO_2000000)
-           log_info("hci_transport_posix: using B1200 for 2 mbps");
-            brate=B1200;
-#endif
-            break;
-        case 3000000:
-#if defined(HAVE_POSIX_B600_MAPPED_TO_3000000)
-            log_info("hci_transport_posix: using B600 for 3 mbps");
-            brate=B600;
-#elif defined(HAVE_POSIX_B2400_MAPPED_TO_3000000)
-            log_info("hci_transport_posix: using B2400 for 3 mbps");
-            brate=B2400;
-#endif
-            break;
-        default:
-            break;
-    }
-    cfsetospeed(&toptions, brate);
-    cfsetispeed(&toptions, brate);
-
-    if( tcsetattr(fd, TCSANOW, &toptions) < 0) {
-        perror("init_serialport: Couldn't set term attributes");
-        return -1;
-    }
-
-    return 0;
+    return btstack_uart_posix_set_baudrate(fd, baudrate);
 }
 
 static void h4_init(const void * transport_config){
