+++ port/posix-h4/Makefile
@@ -1,13 +1,6 @@
 # Makefile for libusb based examples
 BTSTACK_ROOT = ../..
 
-CORE += main.c stdin_support.c
-
-COMMON += \
-	hci_transport_h4_posix.c \
-	btstack_run_loop_posix.c \
-	btstack_link_key_db_fs.c \
-
 CORE += \
 	bluetooth_init_cc2564B_1.2_BT_Spec_4.1.c \
 	btstack_chipset_cc256x.c \
@@ -15,6 +8,12 @@ CORE += \
 	btstack_chipset_em9301.c \
 	btstack_chipset_stlc2500d.c \
 	btstack_chipset_tc3566x.c \
+	btstack_link_key_db_fs.c \
+	btstack_run_loop_posix.c \
+	btstack_uart_posix.c \
+	hci_transport_h4_posix.c \
+	main.c \
+	stdin_support.c \
 #	btstack_chipset_bcm.c \
 
 # TI-WL183x requires TIInit_11.8.32.c
